fix: :bug: 실제 team 과 schema 가 다르게 작성되어 있던 문제 수정
