+++ app/src/api/team/db/team.database.schema.ts
@@ -1,24 +1,53 @@
 import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
 import { HydratedDocument } from 'mongoose';
-import { ScaleTeamBase } from 'src/api/scaleTeam/db/scaleTeam.database.schema';
+import {
+  Flag,
+  ScaleTeamBase,
+} from 'src/api/scaleTeam/db/scaleTeam.database.schema';
 import { UserBase } from 'src/api/user/db/user.database.schema';
+import { TeamBase } from './team.database.base.schema';
 
 export type TeamDocument = HydratedDocument<team>;
 
-// todo: cursus user 완성 후 고치기
 @Schema()
-export class TeamUser extends UserBase {
-  @Prop({ required: true })
-  leader: boolean;
+export class TeamScaleTeam
+  implements Omit<ScaleTeamBase, 'feedbacks' | 'scale'>
+{
+  @Prop()
+  id: number;
 
-  @Prop({ required: true })
-  occurrence: number;
+  @Prop()
+  scaleId: number;
 
-  @Prop({ required: true })
-  validated: boolean;
+  @Prop()
+  createdAt: Date;
 
-  @Prop({ required: true })
-  projectsUserId: number;
+  @Prop()
+  finalMark?: number;
+
+  @Prop()
+  updatedAt: Date;
+
+  @Prop()
+  comment?: string;
+
+  @Prop()
+  beginAt: Date;
+
+  @Prop()
+  correcteds: UserBase[];
+
+  @Prop()
+  corrector: UserBase;
+
+  @Prop()
+  feedback?: string;
+
+  @Prop()
+  filledAt?: Date;
+
+  @Prop()
+  flag: Flag;
 }
 
 @Schema()
@@ -40,64 +69,9 @@ export class TeamsUpload {
 }
 
 @Schema()
-export class TeamBase {
-  @Prop({ required: true })
-  id: number;
-
-  @Prop({ required: true })
-  'closed?': boolean;
-
-  @Prop({ required: true })
-  closedAt?: Date;
-
-  @Prop({ required: true })
-  createdAt: Date;
-
-  @Prop()
-  finalMark?: number;
-
-  @Prop({ required: true })
-  'locked?': boolean;
-
-  @Prop()
-  lockedAt?: Date;
-
-  @Prop({ required: true })
-  name: string;
-
-  @Prop({ required: true })
-  projectId: number;
-
-  @Prop({ required: true })
-  projectSessionId: number;
-
-  @Prop({ required: true })
-  updatedAt: Date;
-
-  @Prop({ required: true })
-  url: string;
-
-  @Prop({ required: true })
-  users: TeamUser[];
-
-  @Prop({ required: true, type: String })
-  status:
-    | 'creating_group'
-    | 'finished'
-    | 'in_progress'
-    | 'waiting_for_correction';
-
-  @Prop()
-  terminatingAt?: Date;
-
-  @Prop()
-  'validated?'?: boolean;
-}
-
-@Schema()
 export class team extends TeamBase {
   @Prop()
-  scaleTeams: ScaleTeamBase[];
+  scaleTeams: TeamScaleTeam[];
 
   @Prop()
   projectGitlabPath: string;
